requirements.txt
numpy
libretro.py
opencv-python

